requirements.txt
claude-agent-sdk>=0.1.44
fastapi>=0.115.0
uvicorn>=0.32.0
fastmcp>=3.0.0
python-dotenv>=1.0.0
